add a few casts to silence warnings
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@1016 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/io/CheckpointWriter.cpp
@@ -48,7 +48,7 @@ void CheckpointWriter::doOutput( ParticleContainer* particleContainer,
 
 		if(_incremental) {
 			/* align file numbers with preceding '0's in the required range from 0 to _numberOfTimesteps. */
-			int num_digits = ceil( log( double( _numberOfTimesteps / _writeFrequency ) ) / log(10.) );
+			int num_digits = (int) ceil( log( double( _numberOfTimesteps / _writeFrequency ) ) / log(10.) );
 			filenamestream << "-" << aligned_number( simstep / _writeFrequency, num_digits, '0' );
 		}
 		filenamestream << ".restart.xdr";
