+++ src/io/VISWriter.cpp
@@ -46,7 +46,7 @@ void VISWriter::doOutput(ParticleContainer* particleContainer,
 
 		if (_incremental) {
 			/* align file numbers with preceding '0's in the required range from 0 to _numberOfTimesteps. */
-			int num_digits = ceil(log(double(_numberOfTimesteps / _writeFrequency)) / log(10.));
+			int num_digits = (int) ceil(log(double(_numberOfTimesteps / _writeFrequency)) / log(10.));
 			filenamestream << aligned_number(simstep / _writeFrequency, num_digits, '0');
 		}
 		filenamestream << ".vis";
